requirements.txt
pyusb>=1.0.0b2
evdev>=0.5.0
